Address pr comments
