+++ src/vivarium_cluster_tools/psimulate/cluster/__init__.py
@@ -14,12 +14,12 @@ from vivarium_cluster_tools.psimulate.cluster.cli_options import (
     with_project,
     with_queue_and_max_runtime,
 )
-from vivarium_cluster_tools.psimulate.cluster.interface import (
+from vivarium_cluster_tools.psimulate.cluster.interface import NativeSpecification
+from vivarium_cluster_tools.psimulate.cluster.validation import (
     AVAILABLE_HARDWARE,
     QUEUE_MAX_RUNTIME_HOURS,
     VALID_PROJECTS,
     VALID_QUEUES,
-    NativeSpecification,
     validate_cluster_environment,
     validate_hardware,
     validate_project,
